+++ man/roxygen/meta.R
@@ -1,7 +1,6 @@
 list(
   rd_family_title = list(
     lookup = "Address Lookup API: ",
-    amenity = "Search amenities: ",
     geocoding = "Geocoding strings: ",
     reverse = "Reverse geocoding coordinates: ",
     spatial = "Get spatial (\\code{sf}) objects: ",