+++ tests/testthat/test-reverse_geo_lite_sf.R
@@ -214,7 +214,7 @@ test_that("Dedupe", {
   lats <- rep(c(40.75728, 55.95335), 50)
   longs <- rep(c(-73.98586, -3.188375), 50)
 
-  expect_silent(dup <- reverse_geo_lite_sf(lats, longs))
+  expect_silent(dup <- reverse_geo_lite_sf(lats, longs, progressbar = FALSE))
 
   expect_s3_class(dup, "sf")
   expect_s3_class(dup, "tbl")
@@ -229,4 +229,23 @@ test_that("Dedupe", {
 
   expect_equal(nrow(dedup), 2)
   expect_equal(as.character(dedup$address), nms)
+})
+
+test_that("Progress bar", {
+  skip_on_cran()
+  skip_if_api_server()
+  skip_if_offline()
+
+  lat <- c(40.75728, 55.95335)
+  long <- c(-73.98586, -3.188375)
+
+  # No pbar
+  expect_silent(reverse_geo_lite_sf(lat[1], long[1]))
+  expect_silent(reverse_geo_lite_sf(lat[1], long[1], progressbar = TRUE))
+
+  # Get a pbar
+  expect_output(aa <- reverse_geo_lite_sf(lat, long), "1/2")
+
+  # Not
+  expect_silent(aa <- reverse_geo_lite_sf(lat, long, progressbar = FALSE))
 })